+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Use latest instead of a version for stableRust in flake.nix
   - `nix flake update`
   - Remove flake-utils dependency from flake.nix
+-
 
 ## v0.5.2 - 2025-10-6
 - Fix documentation